// src/model/ecanConsignmentNote.model.ts
export interface ECanConsignmentNoteReqModel {
    // 郵遞區號
    ForwarderPostalCode: string;
    ForwarderTransportCode: string;
    PrintDate: string;
    ExpectDeliveryArrivalDate: string;
    OuterCode: string;
    TradesOrderGroupCode: string;
    ReceiverAddress: string;
    ReceiverPhone: string;
    ReceiverName: string;
    SenderName: string;
    SenderPhone: string;
    AgencyCharge: string;
    SupplierCode: string;
    Note: string;
    TemperatureTypeDef: string;
    OuterCodeBarcode: string;
    ForwarderPostalCodeBarcode: string;
    SupplierCodeBarcode: string;
    Quantity: string;
    SaleProductSKUOuterId: string;
    SKUPropertySet: string;
    TradesOrderCode: string;
    TradesOrderSlaveCode: string;
    SortType: string;
    SaleProductTitle: string;
}